requirements.txt
pycryptodome==3.19.0
PyYAML==6.0.1
